backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
mysql-connector-python
pydantic
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
